requirements.txt
aiogram>=3.4.1
aiohttp>=3.9.3
python-dotenv>=1.0.0
aiogram-calendar>=0.5.0
apscheduler>=3.10.4


📌 Список идей и функций, которые ты хотел добавить

Я сгруппировал их по смыслу.

🧭 1. Логика поиска и подписок
🔔 Подписка на цену

Ты хотел:

возможность подписаться на маршрут + диапазон дат + максимальную цену,

чтобы бот проверял цены периодически (например, каждые 2 часа),

и присылал уведомление, если цена упала ниже заданного порога.

Формулировка из тебя:

“интервальный поиск например каждые 2 часа, для подписок на конкретные диапазоны и конкретную максимальную стоимость”

🧭 Сдвигаемый диапазон дат с фиксированным интервалом

Ты хотел, чтобы:

если между вылетом и возвращением 21 день — он всегда оставался 21 день,

при поиске можно было двигать обе даты одновременно, но не менять интервал.

“двигать можно только даты отлёта вместе с датами прилета на равное количество дней и в одну и ту же сторону”



✈️ 3. Расширенная информация о перелётах

Ты хотел, чтобы карточки предложений были не просто ценой, а полноценным инструментом принятия решения:

В карточке перелёта:

пересадки (где, сколько времени),

аэропорты вылета и прилёта (полные названия),

авиакомпания,

модель самолёта,

длительность перелёта,

ночной / дневной вылет.

“пересадки, самолёты, аэропорты и т.д. — вся необходимая информация для принятия взвешенного решения”

Ссылки:
✈️ Самолёты

ссылки на сайты производителей моделей самолётов,

фото самолёта,

схема салона (если возможно).

🛫 Аэропорты

ссылки на Википедию или официальный сайт аэропорта,

полезная информация (транзит, ночёвка, терминалы).

“ссылки на источники производителей моделей самолётов, с фотографиями самолётов и планом салона”

🎨 4. UX / UI

Ты очень чётко это сформулировал:

Цена должна быть:

визуально выделена,

вынесена вверх карточки,

быть первым, что бросается в глаза.

“стоимость билетов должна выделяться на фоне другого текста… желательно выносить прайс отдельно от основного текста или поднять его в сообщении”

Также ты хотел:

чтобы интерфейс был современным,

аккуратным,

не перегруженным,

“приятным глазу”.

💳 5. Покупка билетов (позже)

Ты осознанно отложил это:

“Давай покупку отложим на позже, сначала добавим все важные функции”

Но в перспективе:

прямые ссылки на покупку конкретного билета,

привязанные к конкретному офферу.

🛠 6. Архитектурно / продуктово

Ты явно относился к проекту как к продукту:

сначала функциональность,

потом покупка,

потом полировка UX,

потом масштабирование.

“Всё таки продукт сырой. Пока рано.”